+++ Pods/Target Support Files/SwiftTryCatch/SwiftTryCatch-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_SwiftTryCatch : NSObject
-@end
-@implementation PodsDummy_SwiftTryCatch
-@end