+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ## Master
 
 ### Bug fixes
-
+- Fix ReceivedInvocations's type for the method which have only one parameter in AutoMockable.stencil
 - Fix missing folder error that could happen when running a Swift template with existing cache
 - Don't add indentation to empty line when using inline generated code.